refactor(evm): move vaultModals under legacyCard
