+++ apps/evm/src/containers/Vault/VaultCard/Legacy/VaultModals/StakeModal/__tests__/index.spec.tsx
