+++ apps/evm/src/containers/Vault/VaultCard/Legacy/VaultModals/WithdrawFromVestingVaultModal/styles.ts
