--- apps/evm/src/containers/Vault/index.tsx
@@ -1,2 +1,0 @@
-export * from './VaultCard';
-export * from './VaultModals';
